Split some of auth and functions e2e tests
