+++ packages/integration-tests/src/test-e2e/backend_output.test.ts
@@ -21,7 +21,6 @@ import { S3Client } from '@aws-sdk/client-s3';
 import { IAMClient } from '@aws-sdk/client-iam';
 import { DeployedResourcesFinder } from '../find_deployed_resource.js';
 import { DataStorageAuthWithTriggerTestProjectCreator } from '../test-project-setup/data_storage_auth_with_triggers.js';
-import { SQSClient } from '@aws-sdk/client-sqs';
 import { setupDeployedBackendClient } from '../test-project-setup/setup_deployed_backend_client.js';
 import { CloudTrailClient } from '@aws-sdk/client-cloudtrail';
 
@@ -46,7 +45,6 @@ void describe(
     const lambdaClient = new LambdaClient(e2eToolingClientConfig);
     const s3Client = new S3Client(e2eToolingClientConfig);
     const iamClient = new IAMClient(e2eToolingClientConfig);
-    const sqsClient = new SQSClient(e2eToolingClientConfig);
     const resourceFinder = new DeployedResourcesFinder(cfnClient);
     const dataStorageAuthWithTriggerTestProjectCreator =
       new DataStorageAuthWithTriggerTestProjectCreator(
@@ -56,7 +54,6 @@ void describe(
         lambdaClient,
         s3Client,
         iamClient,
-        sqsClient,
         cloudTrailClient,
         resourceFinder
       );
