+++ packages/integration-tests/src/test-e2e/deployment/advanced_auth_and_functions.deployment.test.ts
@@ -1,0 +1,4 @@
+import { defineDeploymentTest } from './deployment.test.template.js';
+import { AdvancedAuthAndFunctionsTestProjectCreator } from '../../test-project-setup/advanced_auth_and_functions.js';
+
+defineDeploymentTest(new AdvancedAuthAndFunctionsTestProjectCreator());
