+++ packages/integration-tests/src/test-e2e/sandbox/advanced_auth_and_functions.sandbox.test.ts
@@ -1,0 +1,4 @@
+import { defineSandboxTest } from './sandbox.test.template.js';
+import { AdvancedAuthAndFunctionsTestProjectCreator } from '../../test-project-setup/advanced_auth_and_functions.js';
+
+defineSandboxTest(new AdvancedAuthAndFunctionsTestProjectCreator());
