+++ packages/integration-tests/src/test-in-memory/data_storage_auth_with_triggers.test.ts
@@ -52,11 +52,6 @@ void it('data storage auth with triggers', () => {
   assertExpectedLogicalIds(templates.defaultNodeFunc, 'AWS::Lambda::Function', [
     'defaultNodeFunctionlambda5C194062',
     'echoFunclambdaE17DCA46',
-    'funcCustomEmailSenderlambda3CCBA9A6',
-    'funcNoMinifylambda91CDF3E0',
-    'funcWithAwsSdklambda5F770AD7',
-    'funcWithSchedulelambda0B6E4271',
-    'funcWithSsmlambda6A8824A1',
     'handler2lambda1B9C7EFF',
     'node16Functionlambda97ECC775',
     'onUploadlambdaA252C959',
