+++ packages/integration-tests/src/test-project-setup/data_storage_auth_with_triggers.ts
@@ -9,11 +9,7 @@ import path from 'path';
 import { TestProjectCreator } from './test_project_creator.js';
 import { DeployedResourcesFinder } from '../find_deployed_resource.js';
 import assert from 'node:assert';
-import {
-  GetFunctionCommand,
-  InvokeCommand,
-  LambdaClient,
-} from '@aws-sdk/client-lambda';
+import { InvokeCommand, LambdaClient } from '@aws-sdk/client-lambda';
 import {
   amplifySharedSecretNameKey,
   createAmplifySharedSecretName,
@@ -21,22 +17,13 @@ import {
 import { HeadBucketCommand, S3Client } from '@aws-sdk/client-s3';
 import { GetRoleCommand, IAMClient } from '@aws-sdk/client-iam';
 import { AmplifyClient } from '@aws-sdk/client-amplify';
-import {
-  DeleteMessageCommand,
-  ReceiveMessageCommand,
-  SQSClient,
-} from '@aws-sdk/client-sqs';
+
 import {
   CloudTrailClient,
   LookupEventsCommand,
 } from '@aws-sdk/client-cloudtrail';
 import { e2eToolingClientConfig } from '../e2e_tooling_client_config.js';
 import isMatch from 'lodash.ismatch';
-import { TextWriter, ZipReader } from '@zip.js/zip.js';
-import {
-  AdminCreateUserCommand,
-  CognitoIdentityProviderClient,
-} from '@aws-sdk/client-cognito-identity-provider';
 
 /**
  * Creates test projects with data, storage, and auth categories.
@@ -66,16 +53,10 @@ export class DataStorageAuthWithTriggerTestProjectCreator
     private readonly iamClient: IAMClient = new IAMClient(
       e2eToolingClientConfig
     ),
-    private readonly sqsClient: SQSClient = new SQSClient(
-      e2eToolingClientConfig
-    ),
     private readonly cloudTrailClient: CloudTrailClient = new CloudTrailClient(
       e2eToolingClientConfig
     ),
-    private readonly resourceFinder: DeployedResourcesFinder = new DeployedResourcesFinder(),
-    private readonly cognitoClient: CognitoIdentityProviderClient = new CognitoIdentityProviderClient(
-      e2eToolingClientConfig
-    )
+    private readonly resourceFinder: DeployedResourcesFinder = new DeployedResourcesFinder()
   ) {}
 
   createProject = async (e2eProjectDir: string): Promise<TestProjectBase> => {
@@ -92,10 +73,8 @@ export class DataStorageAuthWithTriggerTestProjectCreator
       this.lambdaClient,
       this.s3Client,
       this.iamClient,
-      this.sqsClient,
       this.cloudTrailClient,
-      this.resourceFinder,
-      this.cognitoClient
+      this.resourceFinder
     );
     await fs.cp(
       project.sourceProjectAmplifyDirURL,
@@ -114,7 +93,7 @@ export class DataStorageAuthWithTriggerTestProjectCreator
  */
 class DataStorageAuthWithTriggerTestProject extends TestProjectBase {
   // Note that this is pointing to the non-compiled project directory
-  // This allows us to test that we are able to deploy js, cjs, ts, etc without compiling with tsc first
+  // This allows us to test that we are able to deploy js, cjs, ts, etc. without compiling with tsc first
   readonly sourceProjectRootPath =
     '../../src/test-projects/data-storage-auth-with-triggers-ts';
 
@@ -160,10 +139,8 @@ class DataStorageAuthWithTriggerTestProject extends TestProjectBase {
     private readonly lambdaClient: LambdaClient,
     private readonly s3Client: S3Client,
     private readonly iamClient: IAMClient,
-    private readonly sqsClient: SQSClient,
     private readonly cloudTrailClient: CloudTrailClient,
-    private readonly resourceFinder: DeployedResourcesFinder,
-    private readonly cognitoClient: CognitoIdentityProviderClient
+    private readonly resourceFinder: DeployedResourcesFinder
   ) {
     super(
       name,
@@ -250,42 +227,8 @@ class DataStorageAuthWithTriggerTestProject extends TestProjectBase {
       (name) => name.includes('node16Function')
     );
 
-    const funcWithSsm = await this.resourceFinder.findByBackendIdentifier(
-      backendId,
-      'AWS::Lambda::Function',
-      (name) => name.includes('funcWithSsm')
-    );
-
-    const funcWithAwsSdk = await this.resourceFinder.findByBackendIdentifier(
-      backendId,
-      'AWS::Lambda::Function',
-      (name) => name.includes('funcWithAwsSdk')
-    );
-
-    const funcWithSchedule = await this.resourceFinder.findByBackendIdentifier(
-      backendId,
-      'AWS::Lambda::Function',
-      (name) => name.includes('funcWithSchedule')
-    );
-
-    const funcNoMinify = await this.resourceFinder.findByBackendIdentifier(
-      backendId,
-      'AWS::Lambda::Function',
-      (name) => name.includes('funcNoMinify')
-    );
-    const funcCustomEmailSender =
-      await this.resourceFinder.findByBackendIdentifier(
-        backendId,
-        'AWS::Lambda::Function',
-        (name) => name.includes('funcCustomEmailSender')
-      );
-
     assert.equal(defaultNodeLambda.length, 1);
     assert.equal(node16Lambda.length, 1);
-    assert.equal(funcWithSsm.length, 1);
-    assert.equal(funcWithAwsSdk.length, 1);
-    assert.equal(funcWithSchedule.length, 1);
-    assert.equal(funcCustomEmailSender.length, 1);
 
     const expectedResponse = {
       s3TestContent: 'this is some test content',
@@ -296,19 +239,6 @@ class DataStorageAuthWithTriggerTestProject extends TestProjectBase {
 
     await this.checkLambdaResponse(defaultNodeLambda[0], expectedResponse);
     await this.checkLambdaResponse(node16Lambda[0], expectedResponse);
-    await this.checkLambdaResponse(funcWithSsm[0], 'It is working');
-
-    // Custom email sender assertion
-    await this.assertCustomEmailSenderWorks(backendId);
-
-    await this.assertScheduleInvokesFunction(backendId);
-
-    const expectedNoMinifyChunk = [
-      'var handler = async () => {',
-      '  return "No minify";',
-      '};',
-    ].join('\n');
-    await this.checkLambdaCode(funcNoMinify[0], expectedNoMinifyChunk);
 
     const bucketName = await this.resourceFinder.findByBackendIdentifier(
       backendId,
@@ -457,25 +387,6 @@ class DataStorageAuthWithTriggerTestProject extends TestProjectBase {
     assert.deepStrictEqual(responsePayload, expectedResponse);
   };
 
-  private checkLambdaCode = async (
-    lambdaName: string,
-    expectedCode: string
-  ) => {
-    // get the lambda code
-    const response = await this.lambdaClient.send(
-      new GetFunctionCommand({ FunctionName: lambdaName })
-    );
-    const codeUrl = response.Code?.Location;
-    assert(codeUrl !== undefined);
-    const fetchResponse = await fetch(codeUrl);
-    const zipReader = new ZipReader(fetchResponse.body!);
-    const entries = await zipReader.getEntries();
-    const entry = entries.find((entry) => entry.filename.endsWith('index.mjs'));
-    assert(entry !== undefined);
-    const sourceCode = await entry.getData!(new TextWriter());
-    assert(sourceCode.includes(expectedCode));
-  };
-
   private assertExpectedCleanup = async () => {
     await this.waitForBucketDeletion(this.testBucketName);
     await this.assertRolesDoNotExist(this.testRoleNames);
@@ -616,129 +527,5 @@ class DataStorageAuthWithTriggerTestProject extends TestProjectBase {
       }
       throw err;
     }
-  };
-
-  private assertScheduleInvokesFunction = async (
-    backendId: BackendIdentifier
-  ) => {
-    const TIMEOUT_MS = 1000 * 60 * 2; // 2 minutes
-    const startTime = Date.now();
-    let receivedMessageCount = 0;
-
-    const queue = await this.resourceFinder.findByBackendIdentifier(
-      backendId,
-      'AWS::SQS::Queue',
-      (name) => name.includes('testFuncQueue')
-    );
-
-    // wait for schedule to invoke the function one time for it to send a message
-    while (Date.now() - startTime < TIMEOUT_MS) {
-      const response = await this.sqsClient.send(
-        new ReceiveMessageCommand({
-          QueueUrl: queue[0],
-          WaitTimeSeconds: 20,
-          MaxNumberOfMessages: 10,
-        })
-      );
-
-      if (response.Messages) {
-        receivedMessageCount += response.Messages.length;
-
-        // delete messages afterwards
-        for (const message of response.Messages) {
-          await this.sqsClient.send(
-            new DeleteMessageCommand({
-              QueueUrl: queue[0],
-              ReceiptHandle: message.ReceiptHandle,
-            })
-          );
-        }
-      }
-    }
-
-    if (receivedMessageCount === 0) {
-      assert.fail(
-        `The scheduled function failed to invoke and send a message to the queue.`
-      );
-    }
-  };
-
-  private assertCustomEmailSenderWorks = async (
-    backendId: BackendIdentifier
-  ) => {
-    const TIMEOUT_MS = 1000 * 60 * 2; // 2 minutes
-    const startTime = Date.now();
-    const queue = await this.resourceFinder.findByBackendIdentifier(
-      backendId,
-      'AWS::SQS::Queue',
-      (name) => name.includes('customEmailSenderQueue')
-    );
-
-    assert.strictEqual(queue.length, 1, 'Custom email sender queue not found');
-
-    // Trigger an email sending operation
-    await this.triggerEmailSending(backendId);
-
-    // Wait for the SQS message
-    let messageReceived = false;
-    while (Date.now() - startTime < TIMEOUT_MS && !messageReceived) {
-      const response = await this.sqsClient.send(
-        new ReceiveMessageCommand({
-          QueueUrl: queue[0],
-          WaitTimeSeconds: 20,
-        })
-      );
-
-      if (response.Messages && response.Messages.length > 0) {
-        messageReceived = true;
-        // Verify the message content
-        const messageBody = JSON.parse(response.Messages[0].Body || '{}');
-        assert.strictEqual(
-          messageBody.message,
-          'Custom Email Sender is working',
-          'Unexpected message content'
-        );
-
-        // Delete the message
-        await this.sqsClient.send(
-          new DeleteMessageCommand({
-            QueueUrl: queue[0],
-            ReceiptHandle: response.Messages[0].ReceiptHandle!,
-          })
-        );
-      }
-    }
-
-    assert.strictEqual(
-      messageReceived,
-      true,
-      'Custom email sender was not triggered within the timeout period'
-    );
-  };
-
-  private triggerEmailSending = async (backendId: BackendIdentifier) => {
-    const userPoolId = await this.resourceFinder.findByBackendIdentifier(
-      backendId,
-      'AWS::Cognito::UserPool',
-      () => true
-    );
-
-    assert.strictEqual(userPoolId.length, 1, 'User pool not found');
-
-    const username = `testuser_${Date.now()}@example.com`;
-    const password = 'TestPassword123!';
-
-    await this.cognitoClient.send(
-      new AdminCreateUserCommand({
-        UserPoolId: userPoolId[0],
-        Username: username,
-        TemporaryPassword: password,
-        UserAttributes: [
-          { Name: 'email', Value: username },
-          { Name: 'email_verified', Value: 'true' },
-        ],
-      })
-    );
-    // The creation of a new user should trigger the custom email sender
   };
 }
