+++ packages/integration-tests/src/test-projects/advanced-auth-and-functions/amplify/auth/resource.ts
@@ -1,0 +1,15 @@
+import { defineAuth } from '@aws-amplify/backend';
+import { funcCustomEmailSender } from '../function.js';
+
+const customEmailSenderFunction = {
+  handler: funcCustomEmailSender,
+};
+
+export const auth = defineAuth({
+  loginWith: {
+    email: true,
+  },
+  senders: {
+    email: customEmailSenderFunction,
+  },
+});
