+++ packages/integration-tests/src/test-projects/advanced-auth-and-functions/amplify/func-src/handler_custom_email_sender.ts
