+++ packages/integration-tests/src/test-projects/advanced-auth-and-functions/amplify/func-src/handler_no_minify.ts
