+++ packages/integration-tests/src/test-projects/advanced-auth-and-functions/amplify/test_factories.ts
@@ -1,0 +1,17 @@
+import {
+  funcCustomEmailSender,
+  funcNoMinify,
+  funcWithAwsSdk,
+  funcWithSchedule,
+  funcWithSsm,
+} from './function.js';
+import { auth } from './auth/resource.js';
+
+export const authAndFunctions = {
+  auth,
+  funcWithSsm,
+  funcWithAwsSdk,
+  funcWithSchedule,
+  funcNoMinify,
+  funcCustomEmailSender,
+};
