+++ packages/integration-tests/src/test-projects/data-storage-auth-with-triggers-ts/amplify/test_factories.ts
@@ -1,13 +1,5 @@
 import { data } from './data/resource.js';
-import {
-  defaultNodeFunc,
-  funcWithSsm,
-  funcWithAwsSdk,
-  node16Func,
-  funcWithSchedule,
-  funcNoMinify,
-  funcCustomEmailSender,
-} from './function.js';
+import { defaultNodeFunc, node16Func } from './function.js';
 import { storage } from './storage/resource.js';
 import { auth } from './auth/resource.js';
 
@@ -17,9 +9,4 @@ export const dataStorageAuthWithTriggers = {
   defaultNodeFunc,
   data,
   node16Func,
-  funcWithSsm,
-  funcWithAwsSdk,
-  funcWithSchedule,
-  funcNoMinify,
-  funcCustomEmailSender,
 };
